+++ bot/skills/smile_mode.py
@@ -1,7 +1,8 @@
 import logging
 
 from telegram import Update
-from telegram.ext import MessageHandler, Filters, Updater, CallbackContext
+from telegram.ext import MessageHandler, ContextTypes, filters
+from typing_utils import App
 
 from mode import Mode
 
@@ -11,19 +12,20 @@ mode = Mode(mode_name="smile_mode", default=False, pin_info_msg=True)
 
 
 @mode.add
-def add_smile_mode(upd: Updater, handlers_group: int):
+def add_smile_mode(app: App, handlers_group: int):
     """Set up all handler for SmileMode"""
     logger.info("registering smile-mode handlers")
-    dp = upd.dispatcher
-    dp.add_handler(
-        MessageHandler(~Filters.sticker & ~Filters.animation, smile, run_async=True),
-        handlers_group,
+    app.add_handler(
+        MessageHandler(~filters.Sticker.ALL & ~filters.ANIMATION, smile, block=False),
+        group=handlers_group,
     )
 
 
-def smile(update: Update, context: CallbackContext):
+async def smile(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Delete all messages except stickers or GIFs"""
     logger.debug("remove msg: %s", update.effective_message)
-    context.bot.delete_message(
-        update.effective_chat.id, update.effective_message.message_id, 10
+    if update.effective_chat is None or update.effective_message is None:
+        return
+    await context.bot.delete_message(
+        update.effective_chat.id, update.effective_message.message_id
     )
